udpate: we build a hash table, this finishs
first we can retrieve value from hash table, we need the `table, key` to
get the value in the hash table.

then we implement the delete function, we still pass `table, key` to
get the desired result. But to make the hashtable to the continuism, we
add the concept of tombstone, that is, we delete a k/v pair, we add a
tombstone to the place, so we can still go the the position, but we
can't get the meaningful thing, that a tombstone mark.

Then count the length of hash table, we also add the tombstone into
account.

Signed-off-by: aucker <aucker@qq.com>