feat: full-featured hash table with string interning.
At this part, we implement string interning for hash table, which makes
the string comparision very fast. This is a trade-off.

A little overhead when creating strings to intern them V.S. `=` and `==`
operator is much faster. This is worth, for dynamically typed language,
method calls and instance fields are looked up by name at runtime. So
testing a string slow will slow everything.

Signed-off-by: aucker <aucker@qq.com>